Retain smaller pools of biomass for better mass conservation. Lowered threshold from 1 gram per square meter to 1/10th gram per square meter

Integration of UPGM records into operation/management infrastructure. Initial checkin.

Removed all use of h1db1.inc. Check results against subregion and they are identical.

Removed all use of h1hydro.inc. Check results of irrigated run against subregion and they are identical.

Removed all use of h1temp.inc. Check results of irrigated run against subregion and they are identical.

Removed all use of clayomprop.inc, snowprop.inc, heatcap.inc, heatcond.inc. Also moved .for files into modules Checked results of irrigated run against subregion and they are identical.

Removed all use of hpsd.inc and htheta.inc. Checked results of irrigated run against subregion and they are identical.

Removed all use of vapprop.inc and thdbug.inc. Moved .for files into modules. Checked results of irrigated run against subregion and they are identical.

Removed all use of dvolwparam.inc and lsoda.inc. This only affects darcy hydro solutions but should allow parallel processing of darcy in multiple subregions.  Checked results of irrigated run against subregion and they are identical.

Removed all use of wepp_erosion.inc in lib_hydro and lib_wepp. Moved some files into modules. Checked results of irrigated run against subregion and they are identical.

Moved remaining .for files into modules. Checked results of irrigated run against subregion and they are identical.

Removed all use of p1werm.inc and weppconverge.inc. Moved some lib_wepp .for files into modules. Checked results of irrigated run against subregion and they are identical.

Moved remainder of lib_wepp .for files into modules. Checked results of irrigated run against subregion and they are identical.

Moved soil layer utilities into module. Checked results of irrigated run against subregion and they are identical.

Moved remaing lib_mproc files into modules. Checked results of irrigated run against subregion and they are identical.

Moved remaing sci and stir .for files into modules. Checked results of irrigated run against subregion and they are identical.

Moved confidence interval code into modules. Removed unused file. Checked results of run against subregion and they are identical.

Moved report print code into modules. Removed weps_interface_defs file since all refrerenced files are now in modules. Checked results of run against subregion and they are identical.

Initial checkin of UPGM code that builds under cook and runs the test case correctly.

Converted instantiation of upgm from an array to a single pointer. This can now be contained placed in a biomaterial plant srructure in WEPS, allocated when the plant is growing and deallocated when grow is finished.

Hooks added to use UPGM code to run WEPS crops

Using UPGM runs to completion.

Simplified UPGM WEPS to only use one stage past initialization.

Implemented phasePars hash_state which holds the parameters required for each individual phase instance of the same phase type. When moving the the next phase, these parameters are associated directly with the phase, so not "resetting" is required. If a phase is reused (regrowth), the phaseState values may need to be reset.

Implemented stageGDD and phase_rel_gdd  as a phaseState hash_state. Theses are initialized on phase creation and only need to be reset if a phase is reentered due to regrowth.

Created P100 in doproc. This initializes an UPGM crop to grow in WEPS. Moved the calcualtion of seasonal heat units based on day to maturity into a plant_setup routine and out of cinit. Having this done when the plant is created allows the possibility of adjusting an UPGM crop to grow in a specified number of days at the location specified. Full XML specification needs to be done next.

Disaggegating in process. Extracting partitioning, frost damage and senescence, grain fraction from growth. Incomplete.

Partitioning for each component is now computed external to the growth subroutine. Freeze damage is an independent process. Senescence factor and grain fraction factor are also external from growth.

Results are same for a summer annual crop.

check for zero denominator in cut process height mass weighting

Resolved inconsistencies in shoot.out between WEPS and UPGM crop growth. WEPS was skipping output on days when no growth occurred. It now writes a line for that day.

reenabled setting of dayap

Pulled freeze damage out of growth so it occures before shootgrow. No passing leaf and stem area trend into UPGM ShootGrow. 
Move all cinit code into plant_setup. Removed all remaining cinit code. Corrected scan over plants for wsf value under zero stress.

Winter Annual now grows the same in WEPS crop and UPGM.

UPGM final removed due to issues with Intel compiler. Destructor now called explicitly.

Fixed hydro to prevent divide by zero when no living leaf area on freshly harvested perennial.

Perennial multiple harvested crop, alfalfa now grows the same between UPGM and WEPS crop growth model. all output files are identical.

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@17490 77af8a48-1994-4e4a-b6f4-b1c89642997b

+++ src/lib_crop/crop_data_struct_defs.f95
@@ -47,7 +47,7 @@ module crop_data_struct_defs
     real :: zht    ! Crop height (m)
     real :: dstm   ! Number of crop stems per unit area (#/m^2)
                    ! It is computed by taking the tillering factor times the plant population density.
-    real :: xstmrep   ! a representative diameter so that acdstm*acxstmrep*aczht=acrsai
+    double precision :: xstmrep   ! a representative diameter so that acdstm*acxstmrep*aczht=acrsai
     real :: zrtd      ! Crop root depth (m)
     real :: grainf    ! internally computed grain fraction of reproductive mass
   end type crop_residue
